[CodeClean] cleanup code
Code clean,
1. define macro for internal custom event.
2. cast nnfw type for custom definition.
3. update line comment.

+++ c/src/ml-api-inference-single.c
@@ -2099,7 +2099,7 @@ _ml_validate_model_file (const char *const *model,
   }
 
   /** @todo Make sure num_models is correct for each nnfw type */
-  switch (*nnfw) {
+  switch ((int) *nnfw) {
     case ML_NNFW_TYPE_NNFW:
     case ML_NNFW_TYPE_TVM:
     case ML_NNFW_TYPE_ONNX_RUNTIME:
